Glade to C code generation
When supplied by an XML file produced using GLADE this program will produce a number of C files and a CMakelIst.txt file. If cmake is run followed by make, the aim is to compille and link to form a runnable program with a GTK user intrface. Skeleton subs of the event handlers are included in the generated code, and the programmer must add his/her own code to make the program do something useful.